% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconomiaUNAM-package.R
\docType{package}
\name{EconomiaUNAM-package}
\alias{EconomiaUNAM}
\alias{EconomiaUNAM-package}
\title{EconomiaUNAM: Tutoriales para aprender R en la FE UNAM}
\description{
El objetivo del paquete EconomiaUNAM es: 
    Desarrollar tutoriales en español para enseñar R 
    en las clases de la FE UNAM,
    Ofrecer funciones que simplifiquen la enseñanza de temas
    complicados de matemáticas, estadística y econometría,
    Tener un conjunto de herramientas que faciliten el aprendizaje
    de R para todos los estudiantes de la Facultad de Economía
}
\author{
\strong{Maintainer}: Cesár Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Daniel Amieva
  \item Jorge Hernández
}

Other contributors:
\itemize{
  \item Vianney Wickham [contributor]
}

}
\keyword{internal}
