% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconomiaUNAM-package.R
\docType{package}
\name{EconomiaUNAM-package}
\alias{EconomiaUNAM}
\alias{EconomiaUNAM-package}
\title{EconomiaUNAM: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
